changed id type to Long to see if this resolves the error with adding a new student

+++ java_mysql_curd_api/src/main/java/com/backend/java_mysql_curd_api/controller/StudentController.java
@@ -22,7 +22,7 @@ public class StudentController{
 
     // END POINT TO FETCH SINGLE STUDENT
     @GetMapping(value = "/findStudent/{id}")
-    public Optional<Student> findStudentById(@PathVariable(value = "id") int student_id){
+    public Optional<Student> findStudentById(@PathVariable(value = "id") Long student_id){
         return studentService.getStudentById(student_id);
     }
