+++ java_mysql_curd_api/src/main/java/com/backend/java_mysql_curd_api/repository/StudentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {//was <Student, Long>
+public interface StudentRepository extends JpaRepository<Student, Long> {//was <Student, Long>
     //Long changed to Integer to match data type in the DB
 
 }
